Rename the src models module to types
Pure rename: src/models{,.rs} -> src/types{,.rs} and all
crate::models::/mod models references updated to types. No logic change.
(The LLM-model GGUF dir at chatbot/models/ is unrelated and untouched.)
