+++ chatbot/src/services/discord.rs
@@ -3,7 +3,7 @@ use regex::Regex;
 use serenity::{async_trait, model::channel::Message as DMessage, prelude::*};
 
 use crate::{
-    models::user::{UserAction, UserChannel, UserId},
+    types::user::{UserAction, UserChannel, UserId},
     state_machines::user_state_machine::USER_STATE_MACHINE,
 };
