+++ chatbot/src/types.rs
