+++ chatbot/src/types/bot.rs
